+++ packages/server/src/api/controllers/public/mapping/index.ts
@@ -1,9 +1,9 @@
-import tables from "./tables"
-import applications from "./applications"
-import users from "./users"
-import rows from "./rows"
 import queries from "./queries"
+import rows from "./rows"
+import tables from "./tables"
+import users from "./users"
 import views from "./views"
+import applications from "./workspaces"
 
 export default {
   ...tables,
